Revert footstep trace to kismet version
There were compile issues in the change to reduce allocations in
footstep traces. Revert back to the kismet version, but use MeshOwner
as the world context object so that bIgnoreSelf can ignore it.  That
way MeshOwner->Children can be used directly as the rest of the ignore
list, avoiding the allocations.